Reactor:Restruct FrontPage Home Controller into Different Controllers
