--- WebApplication1/Controllers/MembersController.cs
@@ -1,215 +1,0 @@
-﻿using Konscious.Security.Cryptography;
-using System;
-using System.Collections.Generic;
-using System.Data;
-using System.Data.Entity;
-using System.Linq;
-using System.Net;
-using System.Security;
-using System.Security.Cryptography;
-using System.Text;
-using System.Web;
-using System.Web.Mvc;
-using WebApplication1.Migrations;
-using WebApplication1.Models;
-
-namespace WebApplication1.Controllers
-{
-    public class MembersController : Controller
-    {
-        private DbModel db = new DbModel();
-
-        // GET: Members
-        public ActionResult Index()
-        {
-            return View(db.Member.ToList());
-        }
-
-        // GET: Members/Details/5
-        public ActionResult Details(int? id)
-        {
-            if (id == null)
-            {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
-            Member member = db.Member.Find(id);
-            if (member == null)
-            {
-                return HttpNotFound();
-            }
-            return View(member);
-        }
-
-        // GET: Members/Create
-        public ActionResult Create()
-        {
-            List<Permission> PremmsionList = db.Permission.ToList();
-            var firstLevelNodes = PremmsionList.Where(x => x.RecursiveId == null);
-            StringBuilder dataString = new StringBuilder("[");
-            foreach (var firstLevelNode in firstLevelNodes)  //child is pression
-            {
-                GetTreeViewDataString(firstLevelNode, dataString);
-                dataString.Append(",");
-            }
-            dataString.Append("]");
-            ViewBag.TreeViewData = dataString.ToString();
-            return View();
-        }
-
-        private void GetTreeViewDataString(Permission node, StringBuilder dataString)
-        {
-            dataString.Append($@"{{ 'id': '{node.Value}',  'text': '{node.Title}'");
-            if (node.ChildTable.Count > 0)
-            {
-                dataString.Append(",'children':[");
-                foreach (Permission childNode in node.ChildTable)
-                {
-                    GetTreeViewDataString(childNode, dataString);
-                    dataString.Append(",");
-                }
-                dataString.Append(']');
-            }
-            dataString.Append('}');
-        }
-
-
-        // POST: Members/Create
-        // 若要避免過量張貼攻擊，請啟用您要繫結的特定屬性。
-        // 如需詳細資料，請參閱 https://go.microsoft.com/fwlink/?LinkId=317598。
-        //[HttpPost]
-        //[ValidateAntiForgeryToken]
-        //public ActionResult Create(MemberCreateViewModel memberCreateViewModel)
-        //{
-        //    if (ModelState.IsValid)
-        //    {
-        //        if (db.Member.Any(x => x.Account == memberCreateViewModel.Account))
-        //        {
-        //            ModelState.AddModelError("Account", "帳號重複，請重新輸入");
-        //            return View(memberCreateViewModel);
-        //        }
-        //        else 
-        //        {
-        //            Member member = new Member();
-
-        //            string password = memberCreateViewModel.Password;
-        //            byte[] salt = CreateSalt();
-        //            string saltStr = Convert.ToBase64String(salt); //將 byte 改回字串存回資料表
-        //            byte[] hash = HashPassword(password, salt);
-        //            string hashPasswordStr = Convert.ToBase64String(hash);
-
-        //            member.Account = memberCreateViewModel.Account;
-        //            member.Password = hashPasswordStr;
-        //            member.Salt=saltStr;
-        //            member.NickName=memberCreateViewModel.NickName;
-        //            member.Guid=Guid.NewGuid();
-        //            member.Permission = memberCreateViewModel.Permission;
-
-        //            db.Member.Add(member);
-        //            db.SaveChanges();
-        //            return RedirectToAction("Index");
-        //        }
-        //    }
-
-        //    return View(memberCreateViewModel);
-        //}
-
-        // GET: Members/Edit/5
-        public ActionResult Edit(int? id)
-        {
-            if (id == null)
-            {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
-            Member member = db.Member.Find(id);
-            if (member == null)
-            {
-                return HttpNotFound();
-            }
-            return View(member);
-        }
-
-        // POST: Members/Edit/5
-        // 若要避免過量張貼攻擊，請啟用您要繫結的特定屬性。
-        // 如需詳細資料，請參閱 https://go.microsoft.com/fwlink/?LinkId=317598。
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Include = "Id,Account,Password,Salt,NickName,Guid,Permission,CreateTime,UpdateTime")] Member member)
-        {
-            if (ModelState.IsValid)
-            {
-                db.Entry(member).State = EntityState.Modified;
-                db.SaveChanges();
-                return RedirectToAction("Index");
-            }
-            return View(member);
-        }
-
-        // GET: Members/Delete/5
-        public ActionResult Delete(int? id)
-        {
-            if (id == null)
-            {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
-            Member member = db.Member.Find(id);
-            if (member == null)
-            {
-                return HttpNotFound();
-            }
-            return View(member);
-        }
-
-        // POST: Members/Delete/5
-        [HttpPost, ActionName("Delete")]
-        [ValidateAntiForgeryToken]
-        public ActionResult DeleteConfirmed(int id)
-        {
-            Member member = db.Member.Find(id);
-            db.Member.Remove(member);
-            db.SaveChanges();
-            return RedirectToAction("Index");
-        }
-
-        protected override void Dispose(bool disposing)
-        {
-            if (disposing)
-            {
-                db.Dispose();
-            }
-            base.Dispose(disposing);
-        }
-
-        // Argon2 加密
-        //產生 Salt 功能
-        //使用加密安全隨機數產生器 ( ) 產生隨機鹽
-        private byte[] CreateSalt()
-        {
-            //建立一個大小為 16 ( ) 的位元組陣列buffer儲存產生的 salt
-            var buffer = new byte[16];
-            var rng = new RNGCryptoServiceProvider();
-            rng.GetBytes(buffer);
-            return buffer;
-        }
-
-        // Hash 處理加鹽的密碼功能
-        //將使用者的密碼和產生的鹽作為輸入，使用 Argon2 演算法執行密碼雜湊
-        private byte[] HashPassword(string password, byte[] salt)
-        {
-            var argon2 = new Argon2id(Encoding.UTF8.GetBytes(password));
-
-            //底下這些數字會影響運算時間，而且驗證時要用一樣的值
-            //設定之前生成的鹽
-            argon2.Salt = salt;
-            argon2.DegreeOfParallelism = 8; // 4 核心就設成 8
-            argon2.Iterations = 2; // 迭代運算次數，更高的迭代次數可以提高安全性
-            argon2.MemorySize = 1024; // 1 GB，定義演算法要使用的記憶體大小（以位元組為單位）
-
-            //Argon2 演算法產生 16 位元組雜湊並將其作為位元組數組傳回
-            return argon2.GetBytes(16);
-        }
-
-
-
-
-    }
-}
