--- WebApplication1/Views/Home/Business.cshtml
@@ -1,10 +1,0 @@
-﻿@model string
-
-@{
-    //ViewBag.Title = "Job";
-    //ViewBag.PageTitle = "協會業務";
-    Layout = "~/Views/Shared/_LayoutPage.cshtml";
-}
-
-<h2>Job</h2>
-@Html.Raw(@Model)
