--- WebApplication1/Views/Members/Details.cshtml
@@ -1,82 +1,0 @@
-﻿@model WebApplication1.Models.Member
-
-@{
-    ViewBag.Title = "Details";
-}
-
-<h2>Details</h2>
-
-<div>
-    <h4>Member</h4>
-    <hr />
-    <dl class="dl-horizontal">
-        <dt>
-            @Html.DisplayNameFor(model => model.Account)
-        </dt>
-
-        <dd>
-            @Html.DisplayFor(model => model.Account)
-        </dd>
-
-        <dt>
-            @Html.DisplayNameFor(model => model.Password)
-        </dt>
-
-        <dd>
-            @Html.DisplayFor(model => model.Password)
-        </dd>
-
-        <dt>
-            @Html.DisplayNameFor(model => model.Salt)
-        </dt>
-
-        <dd>
-            @Html.DisplayFor(model => model.Salt)
-        </dd>
-
-        <dt>
-            @Html.DisplayNameFor(model => model.NickName)
-        </dt>
-
-        <dd>
-            @Html.DisplayFor(model => model.NickName)
-        </dd>
-
-        <dt>
-            @Html.DisplayNameFor(model => model.Guid)
-        </dt>
-
-        <dd>
-            @Html.DisplayFor(model => model.Guid)
-        </dd>
-
-        <dt>
-            @Html.DisplayNameFor(model => model.Permission)
-        </dt>
-
-        <dd>
-            @Html.DisplayFor(model => model.Permission)
-        </dd>
-
-        <dt>
-            @Html.DisplayNameFor(model => model.CreateTime)
-        </dt>
-
-        <dd>
-            @Html.DisplayFor(model => model.CreateTime)
-        </dd>
-
-        <dt>
-            @Html.DisplayNameFor(model => model.UpdateTime)
-        </dt>
-
-        <dd>
-            @Html.DisplayFor(model => model.UpdateTime)
-        </dd>
-
-    </dl>
-</div>
-<p>
-    @Html.ActionLink("Edit", "Edit", new { id = Model.Id }) |
-    @Html.ActionLink("Back to List", "Index")
-</p>
