+++ WebApplication1/Views/News/Index.cshtml
@@ -22,7 +22,7 @@
         <div class="row">
             <div class="col-md-3 col-sm-12 col-xs-12">
                 <div class="post-media post-image">
-                    <a href="/Home/News/@item.Id"><img alt="img" src="~/Uploads/news/@item.PhotoPath" class="img-responsive"></a>
+                    <a href="/News/Index/@item.Id"><img alt="img" src="~/Uploads/news/@item.PhotoPath" class="img-responsive"></a>
                 </div>
             </div>
             <div class="col-md-9 col-sm-12 col-xs-12">
@@ -30,7 +30,7 @@
                     <div class="entry-header">
                         <span class="post-meta-date"><i class="fa fa fa-clock-o"></i> @item.IssueTime.Value.ToString("yyyy/MM/dd")</span>
                         <h3 class="entry-title">
-                            <a href="/Home/News/@item.Id">@item.Title</a>
+                            <a href="/News/Index/@item.Id">@item.Title</a>
                         </h3>
                         @{ var replaceInput = item.NewsCkContent.Replace("<p>", "").Replace("</p>", "").Replace("<br>"," ");
                             var replaceInputLength = replaceInput.Length;
@@ -40,7 +40,7 @@
                     </div>
                     <!-- header end -->
                     <div class="post-footer">
-                        <a href="/Home/News/@item.Id" class="btn btn-primary">more</a>
+                        <a href="/News/Index/@item.Id" class="btn btn-primary">more</a>
                     </div>
                 </div>
                 <!-- post-body end -->
